backend/requirements.txt
channels==4.2.0
Django==5.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
